+++ rust/headless-client/src/linux.rs
@@ -1,6 +1,6 @@
 //! Implementation, Linux-specific
 
-use super::{CliCommon, IpcClientMsg, IpcServerMsg, SignalKind, FIREZONE_GROUP, TOKEN_ENV_KEY};
+use super::{CliCommon, IpcClientMsg, IpcServerMsg, FIREZONE_GROUP, TOKEN_ENV_KEY};
 use anyhow::{bail, Context as _, Result};
 use connlib_client_shared::{file_logger, Callbacks, Sockets};
 use connlib_shared::{
@@ -9,12 +9,11 @@ use connlib_shared::{
     LoginUrl,
 };
 use firezone_cli_utils::setup_global_subscriber;
-use futures::{Future, SinkExt, StreamExt};
+use futures::{SinkExt, StreamExt};
 use std::{
     net::{IpAddr, Ipv4Addr, Ipv6Addr},
     os::unix::fs::PermissionsExt,
     path::{Path, PathBuf},
-    pin::pin,
     str::FromStr,
     task::{Context, Poll},
 };
@@ -61,44 +60,6 @@ pub fn default_token_path() -> PathBuf {
     PathBuf::from("/etc")
         .join(connlib_shared::BUNDLE_ID)
         .join("token")
-}
-
-pub(crate) fn run_debug_ipc_service(cli: CliCommon) -> Result<()> {
-    crate::debug_command_setup()?;
-    let rt = tokio::runtime::Runtime::new()?;
-    rt.block_on(async {
-        let mut ipc_service = pin!(ipc_listen(cli));
-        let mut signals = Signals::new()?;
-
-        std::future::poll_fn(|cx| {
-            match signals.poll(cx) {
-                Poll::Ready(SignalKind::Hangup) => {
-                    tracing::info!("Caught Hangup signal");
-                    return Poll::Ready(Ok(()));
-                }
-                Poll::Ready(SignalKind::Interrupt) => {
-                    tracing::info!("Caught Interrupt signal");
-                    return Poll::Ready(Ok(()));
-                }
-                Poll::Pending => {}
-            }
-
-            match ipc_service.as_mut().poll(cx) {
-                Poll::Ready(Ok(())) => {
-                    return Poll::Ready(Err(anyhow::anyhow!(
-                        "Impossible, ipc_listen can't return Ok"
-                    )));
-                }
-                Poll::Ready(Err(error)) => {
-                    return Poll::Ready(Err(error).context("ipc_listen failed"));
-                }
-                Poll::Pending => {}
-            }
-
-            Poll::Pending
-        })
-        .await
-    })
 }
 
 pub(crate) fn check_token_permissions(path: &Path) -> Result<()> {
@@ -228,7 +189,7 @@ pub fn firezone_group() -> Result<nix::unistd::Group> {
     Ok(group)
 }
 
-async fn ipc_listen(cli: CliCommon) -> Result<()> {
+pub(crate) async fn ipc_listen(cli: CliCommon) -> Result<()> {
     // Remove the socket if a previous run left it there
     let sock_path = sock_path();
     tokio::fs::remove_file(&sock_path).await.ok();
